handle errors in dashbord with using view

+++ Models/ViewModels/WorkerPageViewModel.cs
@@ -5,6 +5,6 @@
         public int TotalWorkers { get; set; }
         public int TotalNewWorkers { get; set; }
         public int TotalActiveWorkers { get; set; }
-        public double AverageWorkersPerService { get; set; }
+        public int AverageWorkersPerService { get; set; }
     }
 }
